# websrc/docs/swagger.yaml
basePath: /
definitions:
...
  models.BasicResponse:
    properties:
      Error:
        type: string
      Result:
        type: string
    type: object
  models.GenFileParams:
    properties:
      checkCSV:
        type: boolean
      checkGIF:
        type: boolean
      checkJSON:
        type: boolean
      checkPNG:
        type: boolean
      checkSQL:
        type: boolean
      checkServerJSON:
        type: boolean
      checkServerSQL:
        type: boolean
      checkTXT:
        type: boolean
      checkXML:
        type: boolean
      checkZIP:
        type: boolean
      sizeCSV:
        type: string
      sizeGIF:
        type: string
      sizeJSON:
        type: string
      sizePNG:
        type: string
      sizeSQL:
        type: string
      sizeServerJSON:
        type: string
      sizeServerSQL:
        type: string
      sizeTXT:
        type: string
      sizeXML:
        type: string
      sizeZIP:
        type: string
    type: object
...
  models.MigrateTask:
    properties:
      dummy:
        $ref: '#/definitions/models.GenFileParams'
      operationId:
        type: string
      sourcePoint:
        $ref: '#/definitions/models.ProviderConfig'
      targetPoint:
        $ref: '#/definitions/models.ProviderConfig'
    type: object
  models.ProviderConfig:
    properties:
      bucket:
        type: string
      databaseId:
        type: string
      databaseName:
        type: string
      endpoint:
        type: string
      host:
        type: string
      password:
        type: string
      path:
        type: string
      port:
        type: string
      profileName:
        type: string
      projectId:
        type: string
      provider:
        type: string
      region:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: contact-to-cloud-barista@googlegroups.com
    name: API Support
    url: http://cloud-barista.github.io
  description: MC-Data-Manager REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MC-Data-Manager REST API
  version: latest
paths:
...
  /migration/gcp/aws:
    post:
      consumes:
      - application/json
      description: Migrate data stored in GCP Cloud Storage to AWS S3.
      operationId: MigrationGCPToS3PostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from GCP to AWS S3
      tags:
      - '[Migrate]'
  /migration/gcp/linux:
    post:
      consumes:
      - application/json
      description: Migrate data stored in GCP Cloud Storage to a Linux-based system.
      operationId: MigrationGCPToLinuxPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from GCP to Linux
      tags:
      - '[Migrate]'
  /migration/gcp/ncp:
    post:
      consumes:
      - application/json
      description: Migrate data stored in GCP Cloud Storage to NCP Object Storage.
      operationId: MigrationGCPToNCPPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from GCP to NCP Object Storage
      tags:
      - '[Migrate]'
  /migration/gcp/windows:
    post:
      consumes:
      - application/json
      description: Migrate data stored in GCP Cloud Storage to a Windows-based system.
      operationId: MigrationGCPToWindowsPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from GCP to Windows
      tags:
      - '[Migrate]'
...
swagger: "2.0"